Split out serverless endpoints to separate service
