+++ mongodbatlas/mongodbatlas.go
@@ -116,6 +116,7 @@ type Client struct {
 	Auditing                            AuditingsService
 	AlertConfigurations                 AlertConfigurationsService
 	PrivateEndpoints                    PrivateEndpointsService
+	PrivateServerlessEndpoints          PrivateServerlessEndpointsService
 	PrivateEndpointsDeprecated          PrivateEndpointsServiceDeprecated
 	X509AuthDBUsers                     X509AuthDBUsersService
 	ContinuousSnapshots                 ContinuousSnapshotsService
@@ -263,6 +264,7 @@ func NewClient(httpClient *http.Client) *Client {
 	c.Auditing = &AuditingsServiceOp{Client: c}
 	c.AlertConfigurations = &AlertConfigurationsServiceOp{Client: c}
 	c.PrivateEndpoints = &PrivateEndpointsServiceOp{Client: c}
+	c.PrivateServerlessEndpoints = &PrivateServerlessEndpointsServiceOp{Client: c}
 	c.PrivateEndpointsDeprecated = &PrivateEndpointsServiceOpDeprecated{Client: c}
 	c.X509AuthDBUsers = &X509AuthDBUsersServiceOp{Client: c}
 	c.ContinuousRestoreJobs = &ContinuousRestoreJobsServiceOp{Client: c}
